medshield_ai/requirements.txt
flask==3.1.2
rdkit-pypi==2022.9.5
torch==2.9.0
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
matplotlib==3.8.4
